Fix: decode uint64 in message (#494)

+++ pkg/indexer/decode/decoder/map.go
@@ -156,11 +156,11 @@ func Uint64(m map[string]any, key string) (uint64, error) {
 	if !ok {
 		return 0, errors.Errorf("can't find key: %s", key)
 	}
-	u, ok := val.(int)
+	u, ok := val.(uint64)
 	if !ok {
 		return 0, errors.Errorf("key '%s' is not a uint64", key)
 	}
-	return uint64(u), nil
+	return u, nil
 }
 
 func BoolFromMap(m map[string]any, key string) (bool, error) {
